Reorganise things into /lib folder
https://github.com/sveltejs/kit/discussions/12643
