+++ website/src/learn-content/punctuation.svx
@@ -6,8 +6,8 @@ order: 4
 ---
 
 <script>
-    import OutlineCardGrid from '$lib/cards/OutlineCardGrid.svelte'
-    import ShorthandPassage from '$lib/ShorthandPassage.svelte'
+    import OutlineCardGrid from '$lib/components/cards/OutlineCardGrid.svelte'
+    import ShorthandPassage from '$lib/components/ShorthandPassage.svelte'
 </script>
 
 A lot of grammar falls by the wayside in shorthand, but a few essentials remain. Full stops are the main ones. In Teeline they look like slashes rather than dots. It's essential that it goes through the line. If it doesn't it could be mistaken for the letter 'r', which sits on the line.
