+++ website/src/lib/scripts/helpers.ts
@@ -1,4 +1,4 @@
-import type { OutlineObject } from '../data/interfaces/interfaces';
+import type { OutlineObject } from '$lib/data/interfaces/interfaces';
 
 export const prettify = (outlineNames: string[]): string => {
 	return outlineNames.toString().replace(/,/g, ' / ');
