+++ website/src/lib/scripts/search.ts
@@ -1,4 +1,4 @@
-import type { OutlineObject } from '../data/interfaces/interfaces';
+import type { OutlineObject } from '$lib/data/interfaces/interfaces';
 import { disemvowelWord } from './disemvowel';
 
 export const filterAndSortOutlines = (outlines: OutlineObject[], searchTerm: string) => {
