+++ website/src/routes/api/outlines/+server.ts
@@ -1,4 +1,4 @@
-import { hydratedData } from '../../../scripts/hydrate-outline-data';
+import { hydratedData } from '$lib/scripts/hydrate-outline-data';
 import { json } from '@sveltejs/kit';
 
 export async function GET() {
