+++ website/src/routes/revise/+page.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
-	import type { OutlineObject } from '../../data/interfaces/interfaces';
-	import FlippingOutlineCard from '../../lib/cards/OutlineCardFlipping.svelte';
-	import Container from '../../lib/outlineSVGs/OutlineSVG.svelte';
-	import Lines from '../../lib/outlineSVGs/Lines.svelte';
-	import Toggle from '../../lib/Toggle.svelte';
-	import { prettify, shuffleOutlines } from '../../scripts/helpers';
-	import { hydratedData } from '../../scripts/hydrate-outline-data';
+	import type { OutlineObject } from '$lib/data/interfaces/interfaces';
+	import FlippingOutlineCard from '$lib/components/cards/OutlineCardFlipping.svelte';
+	import Container from '$lib/components/outlineSVGs/OutlineSVG.svelte';
+	import Lines from '$lib/components/outlineSVGs/Lines.svelte';
+	import Toggle from '$lib/components/Toggle.svelte';
+	import { prettify, shuffleOutlines } from '$lib/scripts/helpers';
+	import { hydratedData } from '$lib/scripts/hydrate-outline-data';
 
 	const shuffledSpecialOutlines: OutlineObject[] = shuffleOutlines(
 		hydratedData.filter((outline) => outline.specialOutlineMeanings.length > 0)
